tests by Yee's suggestion
